hashed out a few options
hashed out options which are to do with package managers:  these can be enabled after git cloning so that bash is not running unnecessary commands.  I hashed out the loadkeys as well as this is only necessary if i mess up my system at install and need to set it to my keymap every time i start bash.  I also added an alias for youtube-dl so that I get better audio upon downloading.